+++ camelot/stream.py
@@ -319,10 +319,10 @@ class Stream:
             table_bbox = {}
             for area in self.table_area:
                 x1, y1, x2, y2 = area.split(",")
-                x1 = int(x1)
-                y1 = int(y1)
-                x2 = int(x2)
-                y2 = int(y2)
+                x1 = float(x1)
+                y1 = float(y1)
+                x2 = float(x2)
+                y2 = float(y2)
                 table_bbox[(x1, y2, x2, y1)] = None
         else:
             table_bbox = {(0, 0, width, height): None}
